Add list of move vectors inside aligned hdf5 file
Add list of move vectors in hdf5 aligned file in order to be able
in future commits to see if some of the images got a spurious
jump after alignment. This can be used afterwards to do a
post-alignment, correcting the jumps with an interpolation of the
moving vectors that got the jump.