Correct stored move vector
The stored move vector stores now the rows for which the image has
been moved in a first place, and then the columns in a second place.

The movement vector corresponds to the amount of switched pixels
between the normalized image (original i image without movement),
and the aligned image (for each image i).