Allow moving the indicated images
Develop method allowing to move the indicated images for the
indicated amount in a move vector.

The move vector indicates the amount of [rows, columns] pixels
by which the image has to be moved.

Overwrite the jumped images by images which jumps another indicated
amount.

The idea is to use the move_images method with the output
of another method that will process and 'interpolate' the moving
vectors in order to remove the jumps.