Remove indication of installing h5py via pip
Remove indication of installing h5py via pip. It was giving
problems when trying to install ctalign via setup.py install.

ctalign could not be executed then. h5py installed in the
machine, but ctalign could not find it.

Without indicating this requirement, ctalign works correctly if
h5py is installed in the machine.